refactor(web): split search.ts into per-function modules (#496)
Complete the extraction started in audit/refactor-search-ts. search.ts
drops from 602 → 14 lines (re-export shim only).

New modules under src/lib/search/:
- coverage.ts — buildCoverage, toCoverage
- embedding.ts — getQueryEmbedding
- filterOpts.ts — FilterOpts type, buildMetadataSnippet
- segments.ts — searchSegments (hybrid + metadata)
- grouped.ts — searchGrouped (metadata + FTS)
- mentions.ts — searchMentions

Public API through @/lib/search is unchanged; all 11 callers keep
working without edits. 198/198 tests pass, typecheck clean.
